requirements.txt
aiofiles==25.1.0
aiogram==3.27.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-types==0.7.0
attrs==26.1.0
cachetools==7.0.5
certifi==2026.2.25
charset-normalizer==3.4.7
cssselect==1.4.0
cssutils==2.11.1
frozenlist==1.8.0
idna==3.11
lxml==6.0.4
magic-filter==1.0.12
more-itertools==11.0.2
multidict==6.7.1
premailer==3.10.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
requests==2.33.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
yagmail==0.15.293
yarl==1.23.0
